qemu: chardev: avoid impossible overflow
In the rare case where int and long long are not the same size,
the multiplication of an int variable and an int constant might
overflow. Cast the constant to long long to avoid this.

Fixes: baa4edfb79d5ee861a08b5ec11416c5c156d8cd2

+++ src/qemu/qemu_chardev.c
@@ -247,7 +247,7 @@ qemuChardevGetBackendProps(const virDomainChrSourceDef *chr,
 
         if (reconnect != -1 &&
             virQEMUCapsGet(qemuCaps, QEMU_CAPS_CHARDEV_RECONNECT_MILISECONDS)) {
-            reconnect_ms = reconnect * 1000;
+            reconnect_ms = reconnect * (long long) 1000;
             reconnect = -1;
         }
